+++ src/Controller/SuppTacheListener.java
@@ -3,8 +3,6 @@ package Controller;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-import Model.MenuItems;
-
 public class SuppTacheListener implements ActionListener
 {
 
@@ -18,7 +16,7 @@ public class SuppTacheListener implements ActionListener
 	public void actionPerformed(ActionEvent e) 
 	{
 
-		MainController.removeTache((int)id);
+		MainController.removeTache(id);
 
 
 	}
